refactor: implement specific doctor registration
- Added DoctorRegistrationController to handle doctor registration requests.
- Created DoctorRegistrationRequest for validating doctor registration data.
- Updated API routes to include a specific route for doctor registration.
- Removed user_type field from User model, migration, and factory as it is no longer needed.
